requirements.txt
PyYAML==6.0.2
pytest==9.0.2
